requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
gunicorn==19.7.1
idna==2.7
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==2.1.0
pycodestyle==2.6.0
pylint==2.5.3
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.23
whitenoise==3.3.1
wrapt==1.12.1
